+++ noodles-cram/src/writer/data_container/compression_header/preservation_map.rs
@@ -111,7 +111,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use noodles_sam::record::data::field::{value::Type, Tag};
+    use noodles_sam::record::data::field::{Tag, Type};
 
     use super::*;
     use crate::data_container::compression_header::preservation_map::tag_ids_dictionary::Key;