+++ noodles-sam/src/reader/record/data/field.rs
@@ -1,11 +1,12 @@
 mod tag;
+mod ty;
 mod value;
 
 pub(crate) use self::value::parse_value;
 
 use std::io;
 
-use self::tag::parse_tag;
+use self::{tag::parse_tag, ty::parse_type};
 use crate::record::data::field::{Tag, Value};
 
 pub(super) fn parse_field(src: &mut &[u8]) -> io::Result<Option<(Tag, Value)>> {
@@ -20,8 +21,7 @@ pub(super) fn parse_field(src: &mut &[u8]) -> io::Result<Option<(Tag, Value)>> {
     };
 
     consume_delimiter(&mut buf)?;
-    let ty =
-        value::parse_type(&mut buf).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    let ty = parse_type(&mut buf).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
     consume_delimiter(&mut buf)?;
     let value = parse_value(&mut buf, ty)?;